doc: add todo comment for community messages provider to implement with global subscription
