+++ lib/app/features/chat/views/pages/chat_main_page/chat_main_page.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
 import 'package:ion/app/features/chat/community/providers/community_join_requests_provider.c.dart';
-import 'package:ion/app/features/chat/community/providers/community_messages_subscriber_provider.c.dart';
+import 'package:ion/app/features/chat/community/providers/community_messages_handler.c.dart';
 import 'package:ion/app/features/chat/providers/conversations_provider.c.dart';
 import 'package:ion/app/features/chat/recent_chats/views/components/recent_chat_skeleton/recent_chat_skeleton.dart';
 import 'package:ion/app/features/chat/recent_chats/views/pages/recent_chats_empty_page/recent_chats_empty_page.dart';
